Deleted 'PaymentActivity'
0720
