+++ app/src/main/res/layout/activity_payment.xml
@@ -1,39 +1,43 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout
+    android:id="@+id/payment_ll_paymentActivity"
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    android:background="@drawable/white_rounded_background"
+    android:orientation="vertical"
     tools:context=".JH.PaymentActivity"
-    android:background="@drawable/white_rounded_background" >
+    android:layout_margin="20dp"
+    >
 
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="비용 정산"
-        android:textSize="20dp"
-        android:textStyle="bold" />
+        android:text="[ 비용 정산 ]"
+        android:textSize="15dp"
+        android:textStyle="bold"
+        android:layout_gravity="center"
+        />
 
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal" >
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal">
 
         <LinearLayout
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            android:layout_weight="3" >
+            android:layout_weight="3"
+            android:orientation="horizontal">
 
             <TextView
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
+                android:paddingRight="10dp"
                 android:text="TOTAL"
                 android:textSize="20dp"
-                android:textStyle="bold"
-                android:paddingRight="10dp" />
+                android:textStyle="bold" />
 
             <TextView
                 android:id="@+id/payment_tv_total"
@@ -45,20 +49,20 @@
 
         </LinearLayout>
 
-            <LinearLayout
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal"
-                android:layout_weight="1" >
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:orientation="horizontal">
 
             <TextView
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
+                android:paddingLeft="10dp"
+                android:paddingRight="10dp"
                 android:text="1인당"
                 android:textSize="20dp"
-                android:textStyle="bold"
-                android:paddingRight="10dp"
-                android:paddingLeft="10dp" />
+                android:textStyle="bold" />
 
             <TextView
                 android:id="@+id/payment_tv_PricePerPerson"
@@ -68,27 +72,26 @@
                 android:textSize="20dp"
                 android:textStyle="bold" />
 
-            </LinearLayout>
-
         </LinearLayout>
+
+    </LinearLayout>
 
     <FrameLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content" >
+        android:layout_height="wrap_content">
 
-    <ListView
-        android:id="@+id/payment_lv_itemlist"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
+        <ListView
+            android:id="@+id/payment_lv_itemlist"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" />
 
         <com.google.android.material.floatingactionbutton.FloatingActionButton
             android:id="@+id/payment_btn_addItem"
-            android:layout_gravity="bottom|end"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_gravity="bottom|end"
             android:layout_margin="20dp"
-            android:src="@drawable/ic_baseline_add_24"
-            />
+            android:src="@drawable/ic_baseline_add_24" />
 
     </FrameLayout>
